ruff: don't optimize out the full Ruff run
This bug was causing a big rush for files being added to Git.  Such a
new file was detected in the new code, so Ruff was run (detecting all
the issues in the whole project).  But since the file was not available
in the old code, Ruff was not run at all.  The set of errors reported
was big (since no errors were reported for old code).